# Speed up segmentation conversion and `_transform_mask()`.