`_segmentation_to_detections()`, `_segmentations_to_polylines()`, and `_transform_mask()` create a new full-sized mask for each object in the mask, which scales badly. In our use case we have images with dimension about 22,000 x 9,000, with tens of thousands of objects in them. Looping through each object and creating a new full-sized mask object in each iteration is extremely slow.

The proposed changes scale exponentially better. In a 10,000 x 10,000 image with 10,000 objects, the old method would have taken almost three hours, whereas the new method took only 8 seconds.

Here are timing comparisons with a mask of shape 10,000 x 10,000 on a MacBook M2 Air.

`_segmentation_to_detections()` and `_segmentations_to_polylines()` - similar numbers for both:
| method  | n objects | time (seconds) |
|--- | --- | ---|
| old | 10   | 11.3   |
| | 100 | 101.2 |
| | 1,000 | 1,022.5 |
| | 10,000 | ~10,000 (skipped) |
| new | 10 | 2.0 |
| | 100 | 2.0 |
| | 1,000 | 2.11 |
| | 10,000 | 8.26 |

`transform_mask()`:
| method  | n objects | time (seconds) |
|--- | --- | ---|
| old | 10   | 0.4   |
| | 100 | 3.5 |
| | 1,000 | 35.7 |
| | 10,000 | 352.2 |
| new | 10 | 2.0 |
| | 100 | 1.9 |
| | 1,000 | 2.0 |
| | 10,000 | 7.4 |

## What changes are proposed in this pull request?

- Implement a new function `_find_slices()` to preprocess a mask and find the extent of each unique object. This function is the key to speeding up the others.

- Use that function to speed up `_segmentations_to_detections()`, `_segmentations_to_polylines()`, and `_transform_mask()`.

- Keep track of offsets and overall segmentation mask size, to ensure the new detections have correct relative bounding boxes and the new polylines have the correct coordinates. This includes switching from using `_get_polygons()` to `skimage.measure.label()` in `_parse_thing_instances()`, which avoids the problem of polygon tolerance being off by one or more pixels.

- Combine the core functionality of `_segmentations_to_detections()` and `_segmentations_to_polylines()` into a single function, `_convert_segmentations()`. This eliminates code duplication and allows future segmentation conversions to reuse this code.

- In `_convert_segmentations()`, convert RGB masks to integer masks, matching the way it is done in `_transform_mask()`. This eliminates a number of if/then branches and makes the rest of the function easier to reason about.

- Add unit tests to ensure the changed functions behave as expected.

## How is this patch tested? If it is not, please explain why.

Updated and added new unit tests to check correctness.

Running times comparisons appear above.

## Release Notes

### Is this a user-facing change that should be mentioned in the release notes?

<!--
Please fill in relevant options below with an "x", or by clicking the checkboxes
after submitting this pull request. Example:
-   [x] Selected option
-->

-   [X] No. You can skip the rest of this section.
-   [ ] Yes. Give a description of this change to be included in the release
        notes for FiftyOne users.

### What areas of FiftyOne does this PR affect?

-   [ ] App: FiftyOne application changes
-   [ ] Build: Build and test infrastructure changes
-   [X] Core: Core `fiftyone` Python library changes
-   [ ] Documentation: FiftyOne documentation changes
-   [ ] Other