@kemaleren 2024-03-21
After digging into `eta` to see exactly how `_mask_to_polygons()` worked, I realized that `round_coords` was a bad idea. There are ways to get the behavior I wanted here (basically to undo the interpolation done by `skimage.measure.find_contours()`, but simple rounding isn't the way to do it. It happened to work for the simple cases in the unit tests, but not in general.